fix: 分时段出租功能 - 专家审查问题修复
3 位专家审查发现并修复的问题：

**CRITICAL-1: 空 Vec 语义澄清**
set_specific_date_schedule 传入空 Vec 时，改为自动清除（回退到每周循环），
避免歧义；同时释放 emit SpecificDateCleared 事件。

**HIGH-1: 模式切换时的时序保护**
set_machine_rental_mode 添加检查：机器处于 Rented 状态时不允许切换，
避免租期内模式变化导致的调度保证被破坏。

**HIGH-2: TimeRange.contains_hour end=24 边界**
使用 u16 做比较避免 u8 溢出，修正边界语义。

**MEDIUM-1: 存储膨胀防护**
- WeeklySchedule: 每天最多 10 个 TimeRange
- SpecificDateSchedule: 日期必须在 [今天-30, 今天+365] 内
- 时段列表要求互不重叠（新增 ranges_are_disjoint 辅助函数）

**MEDIUM-2: Block time 用常量**
rent-machine 用 MILLISECS_PER_BLOCK = 6000 常量，提高代码可读性。

**新增测试（9 个）**：
- empty_vec_on_set_specific_date_falls_back_to_weekly
- cannot_switch_mode_while_rented
- set_weekly_schedule_enforces_max_ranges
- set_weekly_schedule_rejects_overlapping_ranges
- set_specific_date_rejects_far_future
- set_weekly_schedule_overwrites_existing
- rental_exactly_2_hours_allowed
- rental_just_under_2_hours_rejected
- clear_nonexistent_date_is_noop

**测试**: 69/69 全部通过（原 41 + time-slot 28）

+++ pallets/online-profile/src/types/basic.rs
@@ -85,8 +85,10 @@ impl TimeRange {
     }
 
     /// 判断给定小时是否在时段内 [start, end)
+    /// 注意：hour 是 0-23，end_hour 可以是 24（表示到次日 00:00）
+    /// 当 hour 在范围 [start, end) 内返回 true
     pub fn contains_hour(&self, hour: u8) -> bool {
-        hour >= self.start_hour && hour < self.end_hour
+        hour >= self.start_hour && (hour as u16) < (self.end_hour as u16)
     }
 
     /// 判断一段时间 [from_hour, to_hour] 是否完全落在本时段内
